Introducing a PluginManager that loads "on demand"...
